+++ src/asymmetric.ts
@@ -1,6 +1,5 @@
-import { nanoid as NonSecureNanoid } from 'nanoid/non-secure';
-import { nanoid as SecureNanoid } from 'nanoid';
 import NaCl from 'tweetnacl';
+import {IPaddingSetUp } from './types';
 const { box, randomBytes } = NaCl;
 
 import type {
@@ -8,17 +7,17 @@ import type {
   IFastyCryptKeys,
   IFastyCryptPairKeys,
   IFastyCryptUint8PairKeys,
-  IPadding,
   IPaddingSettings
 } from './types';
 import {
+  createPadding,
   EncodingToUint8,
   ObjectToUint8,
-  random_int,
-  StringToUint8,
+  segmentDocument,
+  setupNanoid,
+  Uint8From,
   Uint8ToEncoding,
   Uint8ToObject,
-  Uint8ToString,
 } from './utils';
 
 export default class FastyCryptAsymmetric {
@@ -28,15 +27,12 @@ export default class FastyCryptAsymmetric {
   protected SignerEncodedKeys: IFastyCryptPairKeys;
   protected encoding: IFastyCryptEncoding;
   protected SubjectPublicKey?: Uint8Array;
-  protected maxPaddingLength: number = 22;
-  protected minPaddingLength: number = 0;
-  protected nanoid: typeof NonSecureNanoid | typeof SecureNanoid = NonSecureNanoid;
-
+  protected paddingSetup: IPaddingSetUp;
 
   constructor(
     encoding: IFastyCryptEncoding = 'base64',
     keys?: IFastyCryptKeys,
-    paddingsSettings?: IPaddingSettings
+    paddingSettings?: IPaddingSettings
   ) {
     this.encoding = encoding;
     if (keys) {
@@ -44,13 +40,7 @@ export default class FastyCryptAsymmetric {
     } else {
       this.createKeys();
     }
-    if(paddingsSettings){
-      if(typeof paddingsSettings.maxPaddingLength == 'number')
-        this.maxPaddingLength = paddingsSettings.maxPaddingLength;
-      if(typeof paddingsSettings.minPaddingLength == 'number')
-        this.minPaddingLength = paddingsSettings.minPaddingLength;
-      this.nanoid = paddingsSettings.stronger ? SecureNanoid : NonSecureNanoid;
-    }
+    this.paddingSetup = setupNanoid(paddingSettings);
   }
 
   get Uint8PublicKey() {
@@ -81,18 +71,13 @@ export default class FastyCryptAsymmetric {
 
   set staticSubject(publicKey: string | Uint8Array | null) {
     if (publicKey) {
-      let Uint8PublicKey: Uint8Array;
-      if (typeof publicKey == 'string') {
-        Uint8PublicKey = EncodingToUint8(publicKey, this.encoding);
-      } else {
-        Uint8PublicKey = publicKey;
-      }
+      let Uint8PublicKey: Uint8Array = typeof publicKey == 'string' 
+        ? EncodingToUint8(publicKey, this.encoding)
+        : publicKey;
 
-      if (Uint8PublicKey.length !== box.publicKeyLength) {
+      if (Uint8PublicKey.length !== box.publicKeyLength)
         throw new Error('Invalid Subject Public Key');
-      } else {
-        this.SubjectPublicKey = Uint8PublicKey;
-      }
+      this.SubjectPublicKey = Uint8PublicKey;
     } else {
       this.SubjectPublicKey = undefined;
     }
@@ -107,17 +92,6 @@ export default class FastyCryptAsymmetric {
 
   get keys() {
     return this.SignerEncodedKeys as IFastyCryptPairKeys;
-  }
-
-  protected createPadding(): IPadding {
-    const length: number = random_int(this.minPaddingLength, this.maxPaddingLength);
-    const lengthCode: string = length >= 10 ? length.toString() : '0' + length;
-    const padding: Uint8Array = StringToUint8(this.nanoid(length));
-    return {
-      padding,
-      length,
-      lengthCode: StringToUint8(lengthCode)
-    };
   }
 
   static createKeys(
@@ -136,43 +110,49 @@ export default class FastyCryptAsymmetric {
     return this.SignerEncodedKeys;
   }
 
+  protected set Uint8PairKeys(keys: IFastyCryptUint8PairKeys) {
+    const {publicKey, secretKey} = keys;
+    const {publicKeyLength, secretKeyLength} = box;
+
+    let SignerKeys: IFastyCryptUint8PairKeys;
+    let SignerEncodedKeys: IFastyCryptPairKeys;
+    SignerEncodedKeys = {
+      secretKey: Uint8ToEncoding(secretKey, this.encoding),
+      publicKey: Uint8ToEncoding(publicKey, this.encoding),
+    };
+    SignerKeys = keys as IFastyCryptUint8PairKeys;
+    
+    if(publicKey.length !== publicKeyLength || secretKey.length !== secretKeyLength)
+      throw new Error(
+        'Your keys are invalid, please, use an valid keys or execute "createKeys" method to create a new keys.',
+      );
+
+    this.SignerKeys = SignerKeys;
+    this.SignerEncodedKeys = SignerEncodedKeys;
+  }
+
   set keys(keys: IFastyCryptKeys) {
-    this.useKeys(keys);
+    const {publicKey, secretKey} = keys;
+    if (typeof publicKey === 'string' && typeof secretKey === 'string') {
+      const Uint8Keys: IFastyCryptUint8PairKeys = {
+        secretKey: EncodingToUint8(secretKey, this.encoding),
+        publicKey: EncodingToUint8(publicKey, this.encoding),
+      };
+      this.Uint8PairKeys = Uint8Keys;
+    }
+    if(typeof publicKey === 'object' && typeof secretKey === 'object'){
+      this.Uint8PairKeys = keys as IFastyCryptUint8PairKeys;
+    } 
   }
 
   useKeys(keys: IFastyCryptKeys | string): void {
     if (typeof keys == 'string') {
       this.staticSubject = keys;
-    } else if (typeof keys.publicKey === 'string') {
-      const Uint8Keys = {
-        secretKey: EncodingToUint8(keys.secretKey as string, this.encoding),
-        publicKey: EncodingToUint8(keys.publicKey, this.encoding),
-      };
-
-      if (
-        Uint8Keys.publicKey.length == box.publicKeyLength &&
-        Uint8Keys.secretKey.length === box.secretKeyLength
-      ) {
-        this.SignerEncodedKeys = keys as IFastyCryptPairKeys;
-        this.SignerKeys = Uint8Keys;
-        return;
-      }
-    } else if (
-      keys.secretKey.length == box.secretKeyLength &&
-      keys.publicKey.length == box.publicKeyLength
-    ) {
-      this.SignerKeys = keys as IFastyCryptUint8PairKeys;
-      this.SignerEncodedKeys = {
-        secretKey: Uint8ToEncoding(keys.secretKey as Uint8Array, this.encoding),
-        publicKey: Uint8ToEncoding(keys.publicKey, this.encoding),
-      };
-      return;
+    }else{
+      this.keys = keys;
     }
-
-    throw new Error(
-      'Your keys are invalid, please, use an valid keys or execute "createKeys" method to create a new keys.',
-    );
   }
+
 
   protected encryptToUint8(
     document: any,
@@ -180,7 +160,7 @@ export default class FastyCryptAsymmetric {
     secretKey: Uint8Array,
   ): Uint8Array {
     const nonce = randomBytes(box.nonceLength);
-    const {padding, lengthCode: paddingLength} = this.createPadding();    
+    const {padding, lengthCode: paddingLength} = createPadding(this.paddingSetup);    
     let EncryptedDoc: Uint8Array;
     const DocToEncrypt = Uint8Array.from([
       ...padding,
@@ -191,96 +171,64 @@ export default class FastyCryptAsymmetric {
     return EncryptedDoc;
   }
 
-  protected decryptFromUint8(
-    encryptedDocument: Uint8Array,
-    publicKey: Uint8Array,
-    secretKey: Uint8Array,
-  ): any {
-    const paddingLengthUint8 = encryptedDocument.subarray(0, 2);
-    const paddingLength = parseInt(Uint8ToString(paddingLengthUint8));
-    if (
-      !Number.isNaN(paddingLength) &&
-      encryptedDocument.length > 2 + paddingLength + box.nonceLength
-    ) {
-      const nonce = encryptedDocument.subarray(2, box.nonceLength + 2);
-      const EncryptedDoc = encryptedDocument.subarray(box.nonceLength + 2);
-      const document = box.open(EncryptedDoc, nonce, publicKey, secretKey);
-      return document ? Uint8ToObject(document.subarray(paddingLength)) : null;
-    } else {
-      throw new Error('Invalid document.');
-    }
+  protected decryptFromUint8(document: Uint8Array, publicKey: Uint8Array, secretKey: Uint8Array): any {
+    const {encryptedDoc, nonce, paddingLength} = segmentDocument(document, box.nonceLength);
+    const decryptedDoc = box.open(encryptedDoc, nonce, publicKey, secretKey);
+    return decryptedDoc ? Uint8ToObject(decryptedDoc.subarray(paddingLength)) : null;
   }
 
-  encrypt(document: any, receiverPublicKey?: string): string {
-    if (receiverPublicKey || this.SubjectPublicKey) {
-      const ReceiverPublicKey = receiverPublicKey
-        ? EncodingToUint8(receiverPublicKey, this.encoding)
-        : (this.SubjectPublicKey as Uint8Array);
-      const EncryptedDoc = this.encryptToUint8(
-        document,
-        ReceiverPublicKey,
-        this.Uint8SecretKey,
-      );
-      
-      return Uint8ToEncoding(EncryptedDoc, this.encoding);;
-    } else {
-      throw new Error(
-        "You must specify an receiver public key to encrypt an document. If you don't want to use it, you can use ephemeral method.",
-      );
-    }
+  encrypt(document: any, receiverPublicKey?: string): string { 
+    const error_msg = 'You must specify an receiver public key to encrypt an document. If you don\'t want to use it, you can use ephemeral method.';
+    const receiver = Uint8From(receiverPublicKey??this.SubjectPublicKey, this.encoding, error_msg)   
+    const EncryptedDoc = this.encryptToUint8(
+      document,
+      receiver,
+      this.Uint8SecretKey,
+    );
+    return Uint8ToEncoding(EncryptedDoc, this.encoding);
   }
 
   decrypt(encryptedDocument: string, senderPublicKey?: string): any {
-    if (senderPublicKey || this.SubjectPublicKey) {
-      const publicKey = senderPublicKey
-        ? EncodingToUint8(senderPublicKey, this.encoding)
-        : (this.SubjectPublicKey as Uint8Array);
-      return this.decryptFromUint8(
-        EncodingToUint8(encryptedDocument, this.encoding),
-        publicKey,
-        this.Uint8SecretKey,
-      );
-    } else {
-      throw new Error("You must specify an document's sender public key.");
-    }
+    const error_msg = "You must specify an document's sender public key.";
+    let sender = Uint8From(senderPublicKey??this.SubjectPublicKey, this.encoding, error_msg);
+    return this.decryptFromUint8(
+      EncodingToUint8(encryptedDocument, this.encoding),
+      sender,
+      this.Uint8SecretKey,
+    );
   }
 
   ephemeralEncrypt(document: any, receiverPublicKey?: string): string {
-    if (receiverPublicKey || this.SubjectPublicKey) {
-      const ephemeral = box.keyPair();
-      const publicKey = receiverPublicKey
-        ? EncodingToUint8(receiverPublicKey, this.encoding)
-        : (this.SubjectPublicKey as Uint8Array);
-      const EncryptedDoc = this.encryptToUint8(
-        document,
-        publicKey,
-        ephemeral.secretKey,
-      );
-      const EphemeralEncryptedDoc = Uint8Array.from([
-        ...ephemeral.publicKey,
-        ...EncryptedDoc,
-      ]);
+    const error_msg = 'You must specify the public key of the receiver.';
+    const publicKey = Uint8From(receiverPublicKey??this.SubjectPublicKey, this.encoding, error_msg);
+    const ephemeral = box.keyPair();
+    const EncryptedDoc = this.encryptToUint8(
+      document,
+      publicKey,
+      ephemeral.secretKey,
+    );
+    const EphemeralEncryptedDoc = Uint8Array.from([
+      ...ephemeral.publicKey,
+      ...EncryptedDoc,
+    ]);
 
-      return Uint8ToEncoding(EphemeralEncryptedDoc, this.encoding);
-    } else {
-      throw new Error('You must specify the public key of the receiver.');
-    }
+    return Uint8ToEncoding(EphemeralEncryptedDoc, this.encoding);
   }
 
   ephemeralDecrypt(ephemeralDocument: string): any {
     const Uint8Doc = EncodingToUint8(ephemeralDocument, this.encoding);
+    const minDocumentLength = box.publicKeyLength + 2 + box.nonceLength;
 
-    if (Uint8Doc.length > box.publicKeyLength + 2 + box.nonceLength) {
-      const publicKey = Uint8Doc.subarray(0, box.publicKeyLength);
-      const EncryptedDoc = Uint8Doc.subarray(box.publicKeyLength);
-      const document = this.decryptFromUint8(
-        EncryptedDoc,
-        publicKey,
-        this.Uint8SecretKey,
-      );
-      return document;
-    } else {
+    if (Uint8Doc.length <= minDocumentLength) 
       throw new Error('Invalid document.');
-    }
+
+    const publicKey = Uint8Doc.subarray(0, box.publicKeyLength);
+    const EncryptedDoc = Uint8Doc.subarray(box.publicKeyLength);
+    const document = this.decryptFromUint8(
+      EncryptedDoc,
+      publicKey,
+      this.Uint8SecretKey,
+    );
+    return document;
   }
 }